+++ lib/ts/recipe/passwordless/types.ts
@@ -402,7 +402,6 @@ export type TypePasswordlessEmailDeliveryInput = {
     urlWithLinkCode?: string;
     codeLifetime: number;
     preAuthSessionId: string;
-    userContext: any;
 };
 
 export type TypePasswordlessSmsDeliveryInput = {
@@ -412,5 +411,4 @@ export type TypePasswordlessSmsDeliveryInput = {
     urlWithLinkCode?: string;
     codeLifetime: number;
     preAuthSessionId: string;
-    userContext: any;
 };