+++ lib/ts/recipe/thirdpartypasswordless/emaildelivery/services/smtp/index.ts
@@ -30,7 +30,7 @@ export default class SMTPService implements EmailDeliveryInterface<TypeThirdPart
             host: config.smtpSettings.host,
             port: config.smtpSettings.port,
             auth: {
-                user: config.smtpSettings.from.email,
+                user: config.smtpSettings.authUsername || config.smtpSettings.from.email,
                 pass: config.smtpSettings.password,
             },
             secure: config.smtpSettings.secure,